\name{apply_major}
\alias{apply_major}
\title{the difference between apply maps and apply major is that apply major will only allow one value per gid per variable}
\usage{
  apply_major(data, mapping, enclos = parent.frame())
}
\description{
  the difference between apply maps and apply major is that
  apply major will only allow one value per gid per
  variable
}

