\name{combine_aes}
\alias{combine_aes}
\title{combine_aes combines two uneval objects overwriting values of the second that are defined in the first.}
\usage{
  combine_aes(map1, map2)
}
\arguments{
  \item{map1}{an uneval object}

  \item{map2}{an uneval object}
}
\description{
  combine_aes combines two uneval objects overwriting
  values of the second that are defined in the first.
}
\keyword{internal}

