\name{combine_refs}
\alias{combine_refs}
\title{combine_refs relocates reference objects within a layer. It works exactly like
combine_glyphs but does not rescale the x and y variables for the reference
object.}
\usage{
  combine_refs(., data)
}
\description{
  combine_refs relocates reference objects within a layer.
  It works exactly like combine_glyphs but does not rescale
  the x and y variables for the reference object.
}

