\name{first_name}
\alias{first_name}
\title{retrieves the first name used in an expression. To be used with screening
mappings for plyr or regular computation.}
\usage{
  first_name(expr)
}
\description{
  retrieves the first name used in an expression. To be
  used with screening mappings for plyr or regular
  computation.
}

