\name{geom_ref_point}
\alias{geom_ref_point}
\title{geom_ref_point calls a point geom but ensures that only one point is drawn per
glyph}
\usage{
  geom_ref_point(mapping = NULL, data = NULL, stat =
  "identity", position = "identity", ...)
}
\description{
  geom_ref_point calls a point geom but ensures that only
  one point is drawn per glyph
}

