\name{geom_ref_rect}
\alias{geom_ref_rect}
\title{geom_ref_rect calls a rect geom but ensures that only one box is drawn per
glyph}
\usage{
  geom_ref_rect(mapping = NULL, data = NULL, stat =
  "identity", position = "identity", ...)
}
\description{
  geom_ref_rect calls a rect geom but ensures that only one
  box is drawn per glyph
}

