\name{get_ys}
\alias{get_ys}
\title{get_ys retrieves all mappings that need to be altered for plotting on a new y axis}
\usage{
  get_ys(data)
}
\arguments{
  \item{data}{a data.frame}
}
\description{
  get_ys retrieves all mappings that need to be altered for
  plotting on a new y axis
}
\keyword{internal}

