\name{gglayer_build}
\alias{gglayer_build}
\title{gglayer_build prepares an embedded layer (class gglayer) for plotting}
\usage{
  gglayer_build(layer)

  gglayer_build(layer)
}
\description{
  gglayer_build prepares an embedded layer (class gglayer)
  for plotting

  gglayer_build prepares an embedded layer (class gglayer)
  for plotting
}

