\name{ggplyr}
\alias{ggplyr}
\title{Create a ggplyr object}
\usage{
  ggplyr(ggplot)
}
\arguments{
  \item{ggplot}{a ggplot object}
}
\description{
  ggplyr gives a ggplot object the S4 class ggplyr, see
  \code{\link{ggplyr-class}}. ggplyr denotes ggplot objects
  that contain extra information to be used to apply plyr
  functions when plotting.
}

