\name{glyph_plot}
\alias{glyph_plot}
\title{Create a glyphs object}
\usage{
  glyph_plot(ggplot)
}
\arguments{
  \item{ggplot}{a ggplot object #'}
}
\description{
  glyph_plot gives a ggplot object the S4 class `glyphs',
  see \code{\link{glyphs-class}}. glyphs denotes ggplot
  objects that contain extra information to be used to
  embed subplots when plotting. glyphs objects have
  similar, but different print and build methods than
  ggplot2 objects.
}

