\name{grid}
\alias{grid}
\title{grid takes a vector of numbers, divides it into n bins, and rounds each point to the nearest center of a bin.}
\usage{
  grid(x, npoints = 10, range. = range(x))
}
\description{
  grid takes a vector of numbers, divides it into n bins,
  and rounds each point to the nearest center of a bin.
}

