\name{grid_by}
\alias{grid_by}
\title{grid_by is a second order function: it returns a function that can be used on
a data set to provide a grouping variable.}
\usage{
  grid_by(x, y, x.npoints = 10, y.npoints = 10, x.range. =
  NULL, y.range. = NULL)
}
\description{
  grid_by is a second order function: it returns a function
  that can be used on a data set to provide a grouping
  variable.
}

