\name{null_omit}
\alias{null_omit}
\title{null.omit removes the NULL elements from a list and returns the remaining objects as a more concise list.}
\usage{
  null_omit(lst)
}
\arguments{
  \item{lst}{a list}
}
\description{
  null.omit removes the NULL elements from a list and
  returns the remaining objects as a more concise list.
}
\keyword{internal}

