\name{plyr_stats}
\alias{plyr_stats}
\title{plyr_stats is just like ggplot2's internal method calculate_stats but it splits on combinations of PANEL and GLYPH instead of just PANEL}
\usage{
  plyr_stats(panel, data, layers)
}
\description{
  plyr_stats is just like ggplot2's internal method
  calculate_stats but it splits on combinations of PANEL
  and GLYPH instead of just PANEL
}

