\name{update_aes}
\alias{update_aes}
\title{update_aes builds a mapping to correctly guide a layer that has combined individual subplots into a single metaplot.}
\usage{
  update_aes(gmap, imap)
}
\arguments{
  \item{imap}{The mapping for each individual subplot}

  \item{gmap}{The mapping for the single, combined plot}
}
\description{
  update_aes builds a mapping to correctly guide a layer
  that has combined individual subplots into a single
  metaplot.
}
\keyword{internal}

