\name{glayer-class}
\alias{"$,glayer-method"}
\alias{"$<-,glayer-method"}
\alias{"+,ggplot,glayer-method"}
\alias{"+,glyphs,glayer-method"}
\alias{"[,glayer-method"}
\alias{"[<-,glayer-method"}
\alias{"c,glayer-method"}
\alias{"ggtransform,glayer-method"}
\alias{"ls,glayer-method"}
\alias{"rep,glayer-method"}
\alias{"show,glayer-method"}
\alias{glayer-class}
\title{glayer class}
\description{
  glayers are layers made with glyphmaps methods. They are
  equivalent to the layers made by ggplot2 functions in all
  ways except that they contain extra information that is
  used to divide the data into subplots and locate those
  subplots witihn the layer when plotting.
}

