\name{glyph_plot}
\alias{glyph_plot}
\title{Create a glyphs object}
\usage{
  glyph_plot(ggplot)
}
\arguments{
  \item{ggplot}{a ggplot object}
}
\description{
  glyph_plot gives a ggplot object the S4 class ggplyr, see
  \code{\link{glyphs-class}}. ggplyr denotes ggplot objects
  that contain extra information to be used to apply plyr
  functions when plotting.
}

