\name{grid}
\alias{grid}
\title{grid turns an ordinary layer into a gridded version of itself.}
\usage{
  grid(layer, grid.aes = aes(), x.nbin = 10, y.nbin = 10,
  x_scale = identity, y_scale = identity, width.adjust =
  0.95, height.adjust = 0.95, reference = NULL, ply.aes =
  TRUE, .ref = FALSE)
}
\description{
  grid turns an ordinary layer into a gridded version of
  itself.
}

