\name{group_by}
\alias{group_by}
\title{group_by is a second order function: it returns a function that can be used on a data set to provide a grouping variable.}
\usage{
  group_by(vars)
}
\description{
  group_by is a second order function: it returns a
  function that can be used on a data set to provide a
  grouping variable.
}

