\name{in_place}
\alias{in_place}
\title{Combine subplots without adjusting locations}
\usage{
  in_place(layers, mapping = NULL, position = NULL, ...)

  in_place(layers)
}
\arguments{
  \item{layers}{a list of ggplot2 layer objects. These
  should be \code{\link{proto}} objects that work to create
  a graph when added to ggplot().}

  \item{mapping}{an uneval object created with
  \code{\link{aes}}}

  \item{x_scale}{}

  \item{y_scale}{}

  \item{width}{}

  \item{height}{}

  \item{reference}{}

  \item{position}{}

  \item{...}{}

  \item{layers}{a list of ggplot2 layer objects. These
  should be \code{\link{proto}} objects that work to create
  a graph when added to ggplot().}

  \item{mapping}{an uneval object created with
  \code{\link{aes}}}

  \item{x_scale}{}

  \item{y_scale}{}

  \item{width}{}

  \item{height}{}

  \item{reference}{}

  \item{position}{}

  \item{...}{}
}
\description{
  Combine subplots without adjusting locations

  Combine subplots without adjusting locations
}

