\name{in_place}
\alias{in_place}
\title{Combine subplots without adjusting locations}
\usage{
  in_place(layers)
}
\arguments{
  \item{layers}{a list of ggplot2 layer objects. These
  should be \code{\link{proto}} objects that work to create
  a graph when added to ggplot().}
}
\description{
  Combine subplots without adjusting locations
}

